refactored for usage without injecting mappers
usages should be subclasses of the mapping class they want to leverage.
flattened class hierarchies and added flags to control behaviors.
created a means by which to configure subclasses of EnumerationMapper.
Fixed some issues with some of the conventions not working properly or
not being applied where they should be.